+++ src/calender/Time.h
@@ -9,18 +9,18 @@
 #include <cstdint>
 #include <ostream>
 
-bool isValidTime(uint_fast8_t hour, uint_fast8_t min, uint_fast8_t sec);
+bool isValidTime(uint_fast16_t hour, uint_fast16_t min, uint_fast16_t sec);
 
 
 class Time {
 	private:
-		uint_fast8_t _hour;
-		uint_fast8_t _min;
-		uint_fast8_t _sec;
+		uint_fast16_t _hour;
+		uint_fast16_t _min;
+		uint_fast16_t _sec;
 	public:
 		Time() = default;
 
-		Time(uint_fast8_t hour, uint_fast8_t min, uint_fast8_t sec) {
+		Time(uint_fast16_t hour, uint_fast16_t min, uint_fast16_t sec) {
 			if (isValidTime(hour, min, sec)) {
 				_hour = hour;
 				_min = min;
@@ -30,11 +30,11 @@ class Time {
 			}
 		}
 
-		uint_fast8_t gethour() const {
+		uint_fast16_t gethour() const {
 			return _hour;
 		}
 
-		void sethour(uint_fast8_t hour) {
+		void sethour(uint_fast16_t hour) {
 			if (isValidTime(hour, _min, _sec)) {
 				Time::_hour = hour;
 			} else {
@@ -42,11 +42,11 @@ class Time {
 			}
 		}
 
-		uint_fast8_t getmin() const {
+		uint_fast16_t getmin() const {
 			return _min;
 		}
 
-		void setmin(uint_fast8_t min) {
+		void setmin(uint_fast16_t min) {
 			if (isValidTime(_hour, min, _sec)) {
 				Time::_min = min;
 			} else {
@@ -54,11 +54,11 @@ class Time {
 			}
 		}
 
-		uint_fast8_t getsec() const {
+		uint_fast16_t getsec() const {
 			return _sec;
 		}
 
-		void setsec(uint_fast8_t sec) {
+		void setsec(uint_fast16_t sec) {
 			if (isValidTime(_hour, _min, sec)) {
 				Time::_sec = sec;
 			} else {
